+++ libmemdraw/defont.c
@@ -6,7 +6,7 @@ Memsubfont*
 getmemdefont(void)
 {
 	char *hdr, *p;
-	int n;
+	uintptr n;
 	Fontchar *fc;
 	Memsubfont *f;
 	int ld;
@@ -20,7 +20,7 @@ getmemdefont(void)
 	 * declared as char*, not ulong*.
 	 */
 	p = (char*)defontdata;
-	n = (ulong)p & 3;
+	n = (uintptr)p & 3;
 	if(n != 0){
 		memmove(p+(4-n), p, sizeofdefont-n);
 		p += 4-n;
